gossip-api: Refactor and remove wallet update information
